+++ src/components/Chat/Chat.tsx
@@ -87,6 +87,7 @@ export const Chat = <
 
   const {
     channel,
+    channels,
     closeMobileNav,
     getAppSettings,
     latestMessageDatesByChannels,
@@ -94,6 +95,7 @@ export const Chat = <
     navOpen,
     openMobileNav,
     setActiveChannel,
+    setChannels,
     translators,
   } = useChat({ client, defaultLanguage, i18nInstance, initialNavOpen });
 
@@ -110,6 +112,7 @@ export const Chat = <
 
   const chatContextValue = useCreateChatContext({
     channel,
+    channels,
     channelsQueryState,
     client,
     closeMobileNav,
@@ -120,6 +123,7 @@ export const Chat = <
     navOpen,
     openMobileNav,
     setActiveChannel,
+    setChannels,
     theme,
     themeVersion,
     useImageFlagEmojisOnWindows,