+++ src/components/Chat/hooks/useChannelsQueryState.ts
@@ -2,7 +2,7 @@ import { Dispatch, SetStateAction, useState } from 'react';
 import type { APIErrorResponse, ErrorFromResponse } from 'stream-chat';
 
 type ChannelQueryState =
-  | 'uninitialized' // the initial state before the first channels query is trigerred
+  | 'uninitialized' // the initial state before the first channels query is triggered
   | 'reload' // the initial channels query (loading the first page) is in progress
   | 'load-more' // loading the next page of channels
   | null; // at least one channels page has been loaded and there is no query in progress at the moment