+++ src/components/Chat/hooks/useChat.ts
@@ -36,6 +36,7 @@ export const useChat = <
     userLanguage: 'en',
   });
 
+  const [channels, setChannels] = useState<Array<Channel<StreamChatGenerics>>>([]);
   const [channel, setChannel] = useState<Channel<StreamChatGenerics>>();
   const [mutes, setMutes] = useState<Array<Mute<StreamChatGenerics>>>([]);
   const [navOpen, setNavOpen] = useState(initialNavOpen);
@@ -123,6 +124,7 @@ export const useChat = <
 
   return {
     channel,
+    channels,
     closeMobileNav,
     getAppSettings,
     latestMessageDatesByChannels,
@@ -130,6 +132,7 @@ export const useChat = <
     navOpen,
     openMobileNav,
     setActiveChannel,
+    setChannels,
     translators,
   };
 };